Undergraduate-Projects/NuSkill-Capstone/backend/requirements.txt
flask==3.0.0
werkzeug==3.0.1
pyjwt==2.8.0
passlib==1.7.4
snowflake-connector-python==3.10.1
python-dotenv==1.0.1
